demo/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via pip-tools
django-debug-toolbar==1.8
django-model-utils==3.0.0
django-shortuuidfield==0.1.3
django==1.11.6
djangorestframework==3.7.1
first==2.0.1              # via pip-tools
pip-tools==1.10.1
pytz==2017.2              # via django
shortuuid==0.5.0          # via django-shortuuidfield
six==1.11.0               # via django-shortuuidfield, pip-tools
sqlparse==0.2.4           # via django-debug-toolbar
